pyproject.toml
[project]
name = "cordage"
version = "0.0.1"
authors = [
    { name="Max Ploner", email="cordage@maxploner.de" },
    ]
description = "Small framework for computational research data management"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    ]
dependencies=[
  'docstring_parser',
]

[project.optional-dependencies]
tests = [
  'pytest',
  'pytest-mypy',
  'pytest-isort',
  'pytest-flake8',
  'pyyaml',
  'toml',
  'flake8<5.0.0',
  'flake8-black',
  'types-PyYAML',
  'types-toml',
]

dev = [
  'types-PyYAML',
  'types-toml',
]

[project.urls]
"Homepage" = "https://github.com/plonerma/cordage"
"Bug Tracker" = "https://github.com/plonerma/cordage/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
flake8-max-line-length = 210
addopts = "--flake8 --mypy --isort"

[tool.black]
line_length = 120

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[tool.isort]
profile = "black"
line_length = 120
